Prevent opening non-homepage pages fetching blob endpoint
